Alterações na tela de cadastro de empregados.

+++ ags-models/src/main/java/com/ags/dao/EmployeeDao.java
@@ -3,7 +3,7 @@ package com.ags.dao;
 import com.ags.models.Employee;
 
 
-public class EmployeeDao extends GenericImplDao<Employee, Long> {
+public class EmployeeDao extends GenericImplDao<Employee, Integer> {
 
 	public EmployeeDao() {
 		super( Employee.class );
